Solved LeetCode 3501. Maximize Active Section with Trade II (Hard)!

+++ SegmentTree/RangeSumQueryCF.cpp
@@ -2,28 +2,36 @@
 
 using namespace std;
 
-class Solution {
+class Solution
+{
 public:
-    int sumQuery(int left, int right, vector<long long>& prefixSum) {
-        if (left == 0) return prefixSum[right];
+    long long sumQuery(int left, int right, vector<long long> &prefixSum)
+    {
+        if (left == 0)
+            return prefixSum[right];
         return prefixSum[right] - prefixSum[left - 1];
     }
 };
 
-int main() {
+int main()
+{
     int N, Q;
     cin >> N >> Q;
     vector<int> nums(N);
-    for (int i = 0; i < N; i++) cin >> nums[i];
+    for (int i = 0; i < N; i++)
+        cin >> nums[i];
     vector<vector<int>> queries(Q, vector<int>(2));
-    for (int i = 0; i < Q; i++) {
+    for (int i = 0; i < Q; i++)
+    {
         cin >> queries[i][0] >> queries[i][1];
     }
     vector<long long> prefixSum(N);
     prefixSum[0] = nums[0];
-    for (int i = 1; i < N; i++) prefixSum[i] = prefixSum[i - 1] + nums[i]; 
+    for (int i = 1; i < N; i++)
+        prefixSum[i] = prefixSum[i - 1] + nums[i];
     Solution obj;
-    for (int i = 0; i < Q; i++) {
+    for (int i = 0; i < Q; i++)
+    {
         int first = queries[i][0];
         int second = queries[i][1];
         cout << obj.sumQuery(first - 1, second - 1, prefixSum) << "\n";
